Streamly/requirements.txt
streamlit
transformers
Pillow
python-dotenv
requests
PyPDF2
chromadb
openai